rename: 상수 폴더명 수정
