+++ src/features/calendar/index.ts
@@ -1,2 +1,2 @@
 export { generateDate } from './lib/utils'
-export { DAY_OF_THE_WEEK } from './config/constants'
+export { DAY_OF_THE_WEEK } from './consts/constants'
